fix: refactor type and seperate serializing and call
